# feat: comprehensive client library improvements with validated types and builders