--- a/clients/python/kotadb/__init__.py
+++ b/clients/python/kotadb/__init__.py
@@ -14,6 +14,16 @@ Example usage:
 from .client import KotaDB
 from .exceptions import ConnectionError, KotaDBError, ValidationError
 from .types import Document, QueryResult, SearchResult
+from .validated_types import (
+    ValidatedPath,
+    ValidatedDirectoryPath,
+    ValidatedDocumentId,
+    ValidatedTitle,
+    NonZeroSize,
+    ValidatedTimestamp,
+)
+from .validation import ValidationError as ClientValidationError
+from .builders import DocumentBuilder, QueryBuilder, UpdateBuilder
 
 __version__ = "0.2.0"
 __all__ = [
@@ -24,4 +34,14 @@ __all__ = [
     "QueryResult",
     "SearchResult",
     "ValidationError",
+    "ClientValidationError",
+    "ValidatedPath",
+    "ValidatedDirectoryPath", 
+    "ValidatedDocumentId",
+    "ValidatedTitle",
+    "NonZeroSize",
+    "ValidatedTimestamp",
+    "DocumentBuilder",
+    "QueryBuilder",
+    "UpdateBuilder",
 ]