… clients/python/.coverage
… clients/python/README.md
… clients/python/coverage.xml
… clients/python/examples/README.md
… clients/python/examples/builder_patterns.py
… clients/python/examples/comprehensive_usage.py
… clients/python/examples/integration_test.py
… clients/python/examples/performance_test.py
… clients/python/kotadb/validated_types.py
… clients/python/kotadb/validation.py
… clients/python/tests/test_builders.py
… clients/python/tests/test_validated_types.py
… clients/typescript/examples/README.md
… clients/typescript/examples/basic-usage.ts
… clients/typescript/examples/integration-test.ts